Move `BaseCanvasFactory`, `BaseCMapReaderFactory`, and `BaseStandardFontDataFactory` to their own file
Given that these factories are being used in *different* files, for Browser respectively Node.js implementations, it seems reasonable to move them into their own file instead.
